+++ packages/typedash/src/Number/isEven.test.ts
@@ -1,0 +1,13 @@
+import { isEven } from './isEven'
+
+test.each([
+  [2, true],
+  [1, false],
+  [0, true],
+  [958, true],
+  [877, false],
+  [-2, true],
+  [-1, false],
+])('%p => %p', (data, expected) => {
+  expect(isEven(data)).toEqual(expected)
+})