+++ packages/typedash/src/Number/isOdd.test.ts
@@ -1,0 +1,13 @@
+import { isOdd } from './isOdd'
+
+test.each([
+  [2, false],
+  [1, true],
+  [0, false],
+  [958, false],
+  [877, true],
+  [-2, false],
+  [-1, true],
+])('%p => %p', (data, expected) => {
+  expect(isOdd(data)).toEqual(expected)
+})