+++ packages/typedash/src/Record/prop.ts
@@ -1,9 +1,7 @@
 /* eslint-disable @typescript-eslint/ban-types */
 /**
- * @description:
+ * @description
  * Like `prop` from ramda but manually curried.
- *
- * @see https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/lodash/fp.d.ts#L2979
  *
  * @example
  * pipe(
@@ -11,13 +9,7 @@
  *   R.prop('foo'),
  * ) // 5
  */
-type Magic<TObject, TKey extends keyof TObject> = (
-  object: TObject,
-) => TObject[TKey]
-
 export const prop =
-  <TObject extends object, TKey extends keyof TObject>(
-    key: TKey,
-  ): Magic<TObject, TKey> =>
-  (object) =>
+  <TObject extends object, TKey extends keyof TObject>(key: TKey) =>
+  (object: TObject): TObject[TKey] =>
     object[key]