+++ packages/typedash/src/_tests/fail.ts
@@ -1,8 +1,9 @@
-import { tapLog } from '../function/tapLogData'
-import { R, flow, throwError } from '../index'
+import { tapLogData } from '../function/tapLogData'
+import { R, pipe, throwError } from '../index'
 
-export const fail = flow(
-  R.omit(['constructor', 'toString', 'format']),
-  tapLog,
-  throwError,
-)
+export const fail = <
+  T extends { constructor: unknown; toString: unknown; format: unknown },
+>(
+  x: T,
+) =>
+  pipe(x, R.omit(['constructor', 'toString', 'format']), tapLogData, throwError)