+++ packages/eslint-config-typedash/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @typedash/eslint-config-typedash
 
+## 0.1.0
+
+### Minor Changes
+
+- 9daa873: \_ Remove lodash dependency, upgrade fp-ts-std, add ramda-adjunct predicates
+
+## 0.0.4
+
+### Patch Changes
+
+- 77722a9: 📦 Patch bump for new build
+- Updated dependencies [77722a9]
+  - @typedash/tsconfig@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes