+++ packages/p-memoize/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @typedash/p-memoize
 
+## 7.1.3
+
+### Patch Changes
+
+- 77722a9: 📦 Patch bump for new build
+
 ## 7.1.2
 
 ### Patch Changes